refactor(api): move scorecard serializer
